requirements.txt
absl-py==2.0.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cv==1.0.0
cycler==0.12.1
dataclasses-json==0.6.6
exceptiongroup==1.2.2
Flask==3.0.3
Flask-CORS==3.0.10
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
importlib-resources==6.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.15.0
kiwisolver==1.4.5
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.41
langchain-text-splitters==0.2.4
langsmith==0.1.135
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.8.2
mediapipe==0.10.9
ml-dtypes==0.2.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pandas==2.1.4
pillow==11.0.0
pip==23.2.1
propcache==0.2.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==68.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.36
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow==2.15.0
termcolor==2.5.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.41.2
wrapt==1.14.1
yarl==1.15.3
zipp==3.20.2
